requirements.txt
anthropic>=0.39.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
websockets>=13.0
playwright>=1.40.0
pyyaml>=6.0
pygetwindow>=0.0.9
mss>=9.0.0
Pillow>=10.0.0
